+++ services/line-stat-getter.js
@@ -15,7 +15,7 @@ function LineStatGetter(model, params, opts) {
   }
 
   function getAggregateField() {
-    var fieldName = params['aggregate_field'] || schema.idField;
+    var fieldName = params['aggregate_field'] || '*';
     return schema.name + '.' + fieldName;
   }