+++ services/resource-updater.js
@@ -1,19 +1,31 @@
 'use strict';
-var ResourceGetter = require('./resource-getter');
 var Interface = require('forest-express');
+var ResourceGetter = require('./resource-getter');
+var CompositeKeysManager = require('./composite-keys-manager');
 
-function ResourceUpdater(model, params) {
+function ResourceUpdater(model, params, record) {
   var schema = Interface.Schemas.schemas[model.name];
 
   this.perform = function () {
     var where = {};
-    where[schema.idField] = params[schema.idField];
+    var compositeKeysManager = new CompositeKeysManager(model, schema, record);
+
+    if (schema.isCompositePrimary) {
+      where = compositeKeysManager.getRecordConditions(params.recordId);
+    } else {
+      where[schema.idField] = params.recordId;
+    }
 
     return model
-      .update(params, { where: where, individualHooks: true })
+      .update(record, { where: where, individualHooks: true })
       .then(function () {
+        if (schema.isCompositePrimary) {
+          record.forestCompositePrimary =
+            compositeKeysManager.createCompositePrimary();
+        }
+
         return new ResourceGetter(model, {
-          recordId: params[schema.idField]
+          recordId: params.recordId
         }).perform();
       });
   };